Create trait file to encapsule crud and searchable methods
