+++ src/Peoples/Peoples.php
@@ -4,13 +4,26 @@ declare(strict_types=1);
 
 namespace Kanvas\Guild\Peoples;
 
+use Baka\Contracts\Database\ModelInterface;
 use Kanvas\Guild\Contracts\UserInterface;
 use Kanvas\Guild\Peoples\Models\Peoples as ModelsPeoples;
-use Phalcon\Mvc\Model\ResultsetInterface;
+use Kanvas\Guild\Traits\Searchable as SearchableTrait;
 use Phalcon\Utils\Slug;
 
 class Peoples
 {
+    use SearchableTrait;
+
+    /**
+     * Set Model for traits.
+     *
+     * @return ModelInterface
+     */
+    public static function getModel() : ModelInterface
+    {
+        return new ModelsPeoples();
+    }
+
     /**
      * Create a new organization
      *
@@ -32,48 +45,6 @@ class Peoples
         $organization->saveOrFail($data, $createFields);
 
         return $organization;
-    }
-
-    /**
-     * Get all organizations associated to a company
-     *
-     * @param integer $page
-     * @param integer $limit
-     * @return ResultsetInterface
-     */
-    public static function getAll(UserInterface $user, int $page = 1, int $limit = 10) : ResultsetInterface
-    {
-        $offset = ($page - 1) * $limit;
-
-        $organizations = ModelsPeoples::find([
-            'conditions' => 'companies_id = :company_id: AND is_deleted = 0',
-            'bind' => [
-                'company_id' => $user->currentCompanyId()
-            ],
-            'limit' => $limit,
-            'offset' => $offset
-        ]);
-
-        return $organizations;
-    }
-
-    /**
-     * Get a organization by its id
-     *
-     * @param integer $id
-     * @return ModelsPeoples
-     */
-    public static function getById(int $id, UserInterface $user) : ModelsPeoples
-    {
-        return ModelsPeoples::findFirstOrFail(
-            [
-                'conditions' => 'id = :id: AND companies_id = :companies_id: AND is_deleted = 0',
-                'bind' => [
-                    'id' => $id,
-                    'companies_id' => $user->currentCompanyId(),
-                ]
-            ]
-        );
     }
 
     /**
