+++ src/Rotations/Rotations.php
@@ -4,14 +4,27 @@ declare(strict_types=1);
 
 namespace Kanvas\Guild\Rotations;
 
+use Baka\Contracts\Database\ModelInterface;
 use Kanvas\Guild\Contracts\UserInterface;
 use Kanvas\Guild\Pipelines\Models\Pipelines as ModelsPipelines;
 use Kanvas\Guild\Rotations\Models\Rotations as ModelsRotations;
-use Phalcon\Mvc\Model\ResultsetInterface;
+use Kanvas\Guild\Traits\Searchable as SearchableTrait;
 use Phalcon\Utils\Slug;
 
 class Rotations
 {
+    use SearchableTrait;
+
+    /**
+     * Set Model for traits.
+     *
+     * @return ModelInterface
+     */
+    public static function getModel() : ModelInterface
+    {
+        return new ModelsRotations();
+    }
+
     /**
      * Create a new rotation
      *
@@ -29,48 +42,6 @@ class Rotations
         $rotation->saveOrFail();
 
         return $rotation;
-    }
-
-    /**
-     * Get all rotations associated to a company
-     *
-     * @param integer $page
-     * @param integer $limit
-     * @return ResultsetInterface
-     */
-    public static function getAll(UserInterface $user, int $page = 1, int $limit = 10) : ResultsetInterface
-    {
-        $offset = ($page - 1) * $limit;
-
-        $rotations = ModelsRotations::find([
-            'conditions' => 'companies_id = :company_id: AND is_deleted = 0',
-            'bind' => [
-                'company_id' => $user->currentCompanyId()
-            ],
-            'limit' => $limit,
-            'offset' => $offset
-        ]);
-
-        return $rotations;
-    }
-
-    /**
-     * Get a rotation by its id
-     *
-     * @param integer $id
-     * @return ModelsRotations
-     */
-    public static function getById(int $id, UserInterface $user) : ModelsRotations
-    {
-        return ModelsRotations::findFirstOrFail(
-            [
-                'conditions' => 'id = :id: AND companies_id = :companies_id: AND is_deleted = 0',
-                'bind' => [
-                    'id' => $id,
-                    'companies_id' => $user->currentCompanyId(),
-                ]
-            ]
-        );
     }
 
     /**
